+++ mcp918/src/minebot/minebot/pathfinding/ActionFall.java
@@ -30,7 +30,7 @@ public class ActionFall extends ActionPlaceOrBreak {
         MineBot.lookAtBlock(new BlockPos(to.getX(), to.getY() + 1, to.getZ()), false);
         MineBot.forward = true;
         EntityPlayerSP thePlayer = Minecraft.theMinecraft.thePlayer;
-        BlockPos whereAmI = new BlockPos((int) thePlayer.posX, (int) thePlayer.posY, (int) thePlayer.posZ);
+        BlockPos whereAmI = new BlockPos(thePlayer.posX, thePlayer.posY, thePlayer.posZ);
         if (whereAmI.equals(to)) {
             System.out.println("Done falling to " + to);
             MineBot.forward = false;
